requirements.txt
asgiref==3.5.0
cffi==1.15.0
cryptography==36.0.1
dj-database-url==0.5.0
Django==4.0.1
django-cors-headers==3.11.0
djangorestframework==3.13.1
gunicorn==20.1.0
psycopg2==2.9.3
pycparser==2.21
pytz==2021.3
sqlparse==0.4.2
tzdata==2021.5
whitenoise==5.3.0
